moved UISettings class into ui_settings.py.py
