+++ arm/models/models.py
@@ -19,51 +19,6 @@ import arm.config.config as cfg
 from arm.models.job import Job
 
 
-class UISettings(db.Model):
-    """
-    Class to hold the A.R.M ui settings
-    """
-    id = db.Column(db.Integer, autoincrement=True, primary_key=True)
-    use_icons = db.Column(db.Boolean)
-    save_remote_images = db.Column(db.Boolean)
-    bootstrap_skin = db.Column(db.String(64))
-    language = db.Column(db.String(4))
-    index_refresh = db.Column(db.Integer)
-    database_limit = db.Column(db.Integer)
-    notify_refresh = db.Column(db.Integer)
-
-    def __init__(self, use_icons=None, save_remote_images=None,
-                 bootstrap_skin=None, language=None, index_refresh=None,
-                 database_limit=None, notify_refresh=None):
-        self.use_icons = use_icons
-        self.save_remote_images = save_remote_images
-        self.bootstrap_skin = bootstrap_skin
-        self.language = language
-        self.index_refresh = index_refresh
-        self.database_limit = database_limit
-        self.notify_refresh = notify_refresh
-
-    def __repr__(self):
-        return f'<UISettings {self.id}>'
-
-    def __str__(self):
-        """Returns a string of the object"""
-
-        return_string = self.__class__.__name__ + ": "
-        for attr, value in self.__dict__.items():
-            return_string = return_string + "(" + str(attr) + "=" + str(value) + ") "
-
-        return return_string
-
-    def get_d(self):
-        """ Returns a dict of the object"""
-        return_dict = {}
-        for key, value in self.__dict__.items():
-            if '_sa_instance_state' not in key:
-                return_dict[str(key)] = str(value)
-        return return_dict
-
-
 class Notifications(db.Model):
     """
     Class to hold the A.R.M notifications
